requirements.txt
streamlit>=1.28.0
pinecone-client>=2.2.4
sentence-transformers>=2.2.0,<3.0.0
python-dotenv>=0.19.0
scikit-learn>=1.0.0
torch>=2.0.0
transformers>=4.30.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0